requirements.txt
asgiref==3.4.1
boto3==1.20.22
botocore==1.23.22
certifi==2021.10.8
cloudinary==1.28.0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.18.3
django-colorfield==0.5.0
django-cors-headers==3.10.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jet==1.0.8
django-storages==1.12.3
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
